+++ src/main/java/com/lion/codecatcherbe/domain/bookmark/dto/response/BookMarkInfoRes.java
@@ -1,4 +1,4 @@
-package com.lion.codecatcherbe.domain.bookmark.dto;
+package com.lion.codecatcherbe.domain.bookmark.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
